backend/requirements.txt
fastapi[standard]==0.104.1
uvicorn
python-multipart
pydantic
llama-index-core
llama-index-embeddings-openai
llama-index-vector-stores-postgres
llama-index-llms-openai
python-dotenv
PyPDF2
Pillow
pytesseract
pymupdf
spire.doc
slowapi
psycopg2-binary
sqlalchemy
